+++ src/onegov.gazette/onegov/gazette/models/__init__.py
@@ -2,6 +2,7 @@ from onegov.gazette.models.category import Category
 from onegov.gazette.models.issue import Issue
 from onegov.gazette.models.issue import IssueName
 from onegov.gazette.models.notice import GazetteNotice
+from onegov.gazette.models.notice import GazetteNoticeFile
 from onegov.gazette.models.organization import Organization
 from onegov.gazette.models.organization import OrganizationMove
 from onegov.gazette.models.principal import Principal
@@ -10,6 +11,7 @@ from onegov.gazette.models.principal import Principal
 __all__ = [
     'Category',
     'GazetteNotice',
+    'GazetteNoticeFile',
     'Issue',
     'IssueName',
     'Organization',